Planner issue that leads to rescan of the Motion (#94)
Fix issue with rescan of Motion

In some cases on path creation stage, the planner didn't account, that
the outer relation of a join in case of parametrization by outer
parameter from join above, might be rescanned.

Example of approximate plans that planner was able to produce, which lead to error.
HashJoin plan

Gather Motion
    ->  Nested Loop
            ->  Broadcast Motion
                ->  Seq Scan (we get param here)
            ->  Materialize
                ->  Hash Join
                        ->  Broadcast Motion
                            ->  Foreign Scan
                        ->  Hash
                            ->  Function Scan (we use param here)

MergeJoin plan

Nested Loop Semi Join
   ->  Gather Motion
         ->  Seq Scan
   ->  Materialize
         ->  Merge Left Join
               ->  Gather Motion
                     ->  Sort
                           ->  Seq Scan
               ->  Sort
                     ->  Result
                           Filter (we use param here)
                           ->  Materialize
                                 ->  Gather Motion
                                       ->  Seq Scan

In HashJoin we are getting error when we try to rescan Foreign Scan. We do it
because Seq Scan can produce >1 rows so, because of parametrization
we will be forced to rebuild HashTable and rescan Foreign Scan too.
And in MergeJoin we are getting error when we try to rescan Gather Motion of
outer side of Merge Join, for the same reason.

This patch applies same fix, previously used to address same problem
for NestLoop Join.
After creating a motions in join, we check if we have outer
parameters, which can lead to rescanning of outer path, and
if we do have them and we're not rescanable, we add Materialization
to outer path. We do not apply this logic for inner paths because they
are already rescannable by themselves in some form.